+++ compiler-utils/src/testFixtures/stable/com/squareup/anvil/compiler/internal/testing/KotlinCompilationExtensions.kt
@@ -1,8 +1,0 @@
-package com.squareup.anvil.compiler.internal.testing
-
-import com.tschuchort.compiletesting.KotlinCompilation
-import org.jetbrains.kotlin.compiler.plugin.ComponentRegistrar
-
-internal fun KotlinCompilation.setRegistrars(registrars: List<ComponentRegistrar>) {
-  compilerPlugins = registrars
-}