+++ compiler/src/test/java/com/squareup/anvil/compiler/codegen/BindingModuleGeneratorTest.kt
@@ -4,7 +4,6 @@ import com.google.common.truth.Truth.assertThat
 import com.squareup.anvil.annotations.MergeComponent
 import com.squareup.anvil.annotations.MergeSubcomponent
 import com.squareup.anvil.annotations.compat.MergeModules
-import com.squareup.anvil.compiler.KOTLIN_PREVIEW
 import com.squareup.anvil.compiler.USE_IR
 import com.squareup.anvil.compiler.anvilModule
 import com.squareup.anvil.compiler.assumeIrBackend
@@ -398,11 +397,7 @@ class BindingModuleGeneratorTest(
     ) {
       assertThat(exitCode).isError()
 
-      if (KOTLIN_PREVIEW) {
-        assertThat(messages).contains("Source0.kt:6:11")
-      } else {
-        assertThat(messages).contains("Source0.kt: (6, 11)")
-      }
+      assertThat(messages).contains("Source0.kt:6:11")
       assertThat(messages).contains(
         "Class com.squareup.test.ContributingInterface binds com.squareup.test.ParentInterface, " +
           "but the bound type contains type parameter(s) <T, S>. Type parameters in bindings " +
@@ -434,11 +429,7 @@ class BindingModuleGeneratorTest(
     ) {
       assertThat(exitCode).isError()
 
-      if (KOTLIN_PREVIEW) {
-        assertThat(messages).contains("Source0.kt:6:11")
-      } else {
-        assertThat(messages).contains("Source0.kt: (6, 11)")
-      }
+      assertThat(messages).contains("Source0.kt:6:11")
       assertThat(messages).contains(
         "Class com.squareup.test.ContributingInterface binds com.squareup.test.ParentInterface, " +
           "but the bound type contains type parameter(s) <OutputT>. Type parameters in " +