+++ compiler/src/test/java/com/squareup/anvil/compiler/codegen/ContributesToGeneratorTest.kt
@@ -1,7 +1,6 @@
 package com.squareup.anvil.compiler.codegen
 
 import com.google.common.truth.Truth.assertThat
-import com.squareup.anvil.compiler.KOTLIN_PREVIEW
 import com.squareup.anvil.compiler.assumeIrBackend
 import com.squareup.anvil.compiler.compile
 import com.squareup.anvil.compiler.componentInterface
@@ -299,11 +298,7 @@ class ContributesToGeneratorTest {
     ) {
       assertThat(exitCode).isError()
       // Position to the class.
-      if (KOTLIN_PREVIEW) {
-        assertThat(messages).contains("Source0.kt:6:16")
-      } else {
-        assertThat(messages).contains("Source0.kt: (6, 16)")
-      }
+      assertThat(messages).contains("Source0.kt:6:16")
     }
   }
 
@@ -328,11 +323,7 @@ class ContributesToGeneratorTest {
       ) {
         assertThat(exitCode).isError()
         // Position to the class.
-        if (KOTLIN_PREVIEW) {
-          assertThat(messages).contains("Source0.kt:7")
-        } else {
-          assertThat(messages).contains("Source0.kt: (7, ")
-        }
+        assertThat(messages).contains("Source0.kt:7")
       }
     }
   }
@@ -357,11 +348,7 @@ class ContributesToGeneratorTest {
       ) {
         assertThat(exitCode).isError()
         // Position to the class.
-        if (KOTLIN_PREVIEW) {
-          assertThat(messages).contains("Source0.kt:6")
-        } else {
-          assertThat(messages).contains("Source0.kt: (6, ")
-        }
+        assertThat(messages).contains("Source0.kt:6")
       }
     }
   }