+++ compiler/src/test/java/com/squareup/anvil/compiler/dagger/ProvidesMethodFactoryGeneratorTest.kt
@@ -1,7 +1,6 @@
 package com.squareup.anvil.compiler.dagger
 
 import com.google.common.truth.Truth.assertThat
-import com.squareup.anvil.compiler.KOTLIN_PREVIEW
 import com.squareup.anvil.compiler.USE_IR
 import com.squareup.anvil.compiler.WARNINGS_AS_ERRORS
 import com.squareup.anvil.compiler.anvilModule
@@ -2560,11 +2559,7 @@ public final class DaggerComponentInterface implements ComponentInterface {
       assumeFalse(useDagger)
 
       assertThat(exitCode).isError()
-      if (KOTLIN_PREVIEW) {
-        assertThat(messages).contains("Source0.kt:5:3")
-      } else {
-        assertThat(messages).contains("Source0.kt: (5, 3)")
-      }
+      assertThat(messages).contains("Source0.kt:5:3")
       assertThat(messages).contains(
         "Dagger provider methods must specify the return type explicitly when using Anvil. " +
           "The return type cannot be inferred implicitly."